fix Makefile, docker_rebuild.sh

+++ docker_rebuild.sh
@@ -8,7 +8,10 @@ if [ "$(docker ps -aqf "name=oj_web")" != "" ]; then
 	docker rm -f oj_web
 fi
 
-docker rmi -f oj_web
+if [ "$(docker images -q oj_web)" != "" ]; then
+	docker rmi -f oj_web
+fi
+
 docker build -t oj_web .
 
 docker run -d -p 80:8000 -v $(pwd)/data:/var/nthuoj -v $(pwd)/media:/NTHUOJ_web/media --name=oj_web --cpu-shares 20 --restart=always oj_web
